fix: redo changes as some methods are moved with previous commit
