+++ backend/index.js
@@ -9,4 +9,5 @@ module.exports = {
   ...require('./members-area-methods'), //TODO: remove it once we finish NPM movement
   ...require('./members-data-methods'), //TODO: remove it once we finish NPM movement
   ...require('./cms-data-methods'), //TODO: remove it once we finish NPM movement
+  ...require('./sso-methods'),
 };
