+++ public/consts.js
@@ -83,8 +83,6 @@ const LIGHTBOX_NAMES = {
   CONTACT_US: 'Contact Us',
 };
 
-const ABMP_MEMBERS_HOME_URL = 'https://www.abmp.com/members';
-
 const FREE_WEBSITE_TEXT_STATES = {
   ENABLED: 'This is the default and will auto-populate with the information entered on this page.',
   DISABLED: 'To deactivate, please opt in via Personal Details.',
@@ -103,7 +101,6 @@ module.exports = {
   DEBOUNCE_DELAY,
   MEMBERS_FIELDS,
   LIGHTBOX_NAMES,
-  ABMP_MEMBERS_HOME_URL,
   FREE_WEBSITE_TEXT_STATES,
   DEFAULT_BUSINESS_NAME_TEXT,
   DEFAULT_PROFILE_IMAGE,
